pyproject.toml
[project]
name = "web-agent"
version = "0.1.0"
description = "Browser-based AI coding agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "anthropic>=0.49.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.0",
]
